update timeline claim error and create models
