+++ src/mlflow_monitor/workflow.py
@@ -35,7 +35,8 @@ from mlflow_monitor.errors import (
     PreparedContextConsistencyViolation,
     PrepareStageError,
 )
-from mlflow_monitor.gateway import MonitoringGateway, TimelineState
+from mlflow_monitor.gateway import MonitoringGateway
+from mlflow_monitor.gateway_models import TimelineState
 from mlflow_monitor.invariant import validate_contract_check_result
 from mlflow_monitor.recipe_compiler import CompiledRecipe, EffectiveRecipePlan
 from mlflow_monitor.utils import canonical_json
