+++ tests/test_monitor.py
@@ -22,11 +22,11 @@ from mlflow_monitor.errors import (
     PREPARE_BASELINE_OVERRIDE_EXISTING_BASELINE,
     GatewayConsistencyViolation,
 )
-from mlflow_monitor.gateway import (
+from mlflow_monitor.gateway import InMemoryMonitoringGateway
+from mlflow_monitor.gateway_models import (
     CreateOrReuseMonitoringRunResult,
     GatewayConfig,
     IdempotencyKey,
-    InMemoryMonitoringGateway,
     TimelineState,
 )
 from mlflow_monitor.orchestration import run_orchestration
